+++ .idea/compiler.xml
@@ -7,13 +7,33 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="demo" />
+        <module name="assembly" />
+        <module name="main-app" />
+        <module name="agv" />
+        <module name="product" />
+        <module name="warehouse" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="Agv" target="21" />
+      <module name="Assembly" target="21" />
+      <module name="MainApp" target="21" />
+      <module name="products" target="21" />
+      <module name="Warehouse" target="21" />
+    </bytecodeTargetLevel>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
+      <module name="Agv" options="-parameters" />
+      <module name="Assembly" options="-parameters" />
+      <module name="MainApp" options="-parameters" />
+      <module name="Warehouse" options="-parameters" />
+      <module name="agv" options="-parameters" />
+      <module name="assembly" options="-parameters" />
       <module name="demo" options="-parameters" />
+      <module name="main-app" options="-parameters" />
+      <module name="product" options="-parameters" />
+      <module name="warehouse" options="-parameters" />
     </option>
   </component>
 </project>
