+++ .idea/encodings.xml
@@ -1,6 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/Products/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Products/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/MainApp/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/agv/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/assembly/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/main-app/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/product/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/demo/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/demo/warehouse/src/main/java" charset="UTF-8" />
   </component>
 </project>
