+++ demo/assembly/src/main/resources/application.properties
