+++ demo/product/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module product.module {
+    requires spring.boot;
+    requires spring.boot.autoconfigure;
+    requires spring.web;
+    requires spring.data.jpa;
+    requires com.fasterxml.jackson.databind;
+
+    exports com.example.product;
+}
