+++ demo/src/test/java/com/example/demo/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package src.test.java.com.example.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
