+++ demo/warehouse/src/main/java/com/example/warehouse/WarehouseEndpoint.java
@@ -1,0 +1,41 @@
+package com.example.warehouse;
+
+import com.example.warehouse.WarehouseService;
+import com.example.warehouse.warehouse.GetInventoryResponse;
+import com.example.warehouse.warehouse.InsertItemRequest;
+import com.example.warehouse.warehouse.PickItemRequest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.ws.server.endpoint.annotation.Endpoint;
+import org.springframework.ws.server.endpoint.annotation.PayloadRoot;
+import org.springframework.ws.server.endpoint.annotation.RequestPayload;
+import org.springframework.ws.server.endpoint.annotation.ResponsePayload;
+
+@Endpoint
+@Component
+public class WarehouseEndpoint {
+
+    private final WarehouseService warehouseService;
+
+    @Autowired
+    public WarehouseEndpoint(WarehouseService warehouseService) {
+        this.warehouseService = warehouseService;
+    }
+
+    @PayloadRoot(namespace = "http://example.org/warehouse", localPart = "PickItem")
+    public void pickItem(@RequestPayload PickItemRequest request) {
+        // Logic to request an item from the warehouse
+    }
+
+    @PayloadRoot(namespace = "http://example.org/warehouse", localPart = "InsertItem")
+    public void insertItem(@RequestPayload InsertItemRequest request) {
+        // Logic to insert an item into the warehouse
+    }
+
+    @PayloadRoot(namespace = "http://example.org/warehouse", localPart = "GetInventory")
+    @ResponsePayload
+    public GetInventoryResponse getInventory() {
+        // Logic to get a response with inventory data
+        return null;
+    }
+}
