+++ demo/warehouse/src/main/java/com/example/warehouse/warehouse/GetInventoryResponse.java
@@ -1,0 +1,147 @@
+
+package com.example.warehouse.warehouse;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Java class for anonymous complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType&gt;
+ *   &lt;complexContent&gt;
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+ *       &lt;sequence&gt;
+ *         &lt;element name="Inventory"&gt;
+ *           &lt;complexType&gt;
+ *             &lt;complexContent&gt;
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+ *                 &lt;sequence&gt;
+ *                 &lt;/sequence&gt;
+ *               &lt;/restriction&gt;
+ *             &lt;/complexContent&gt;
+ *           &lt;/complexType&gt;
+ *         &lt;/element&gt;
+ *         &lt;element name="State" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
+ *         &lt;element name="TimeStamp" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
+ *       &lt;/sequence&gt;
+ *     &lt;/restriction&gt;
+ *   &lt;/complexContent&gt;
+ * &lt;/complexType&gt;
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+    "inventory",
+    "state",
+    "timeStamp"
+})
+@XmlRootElement(name = "GetInventory", namespace = "http://example.org/warehouse")
+public class GetInventoryResponse {
+
+    @XmlElement(name = "Inventory", namespace = "http://example.org/warehouse", required = true)
+    protected GetInventoryResponse.Inventory inventory;
+    @XmlElement(name = "State", namespace = "http://example.org/warehouse")
+    protected int state;
+    @XmlElement(name = "TimeStamp", namespace = "http://example.org/warehouse", required = true)
+    protected String timeStamp;
+
+    /**
+     * Gets the value of the inventory property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link GetInventoryResponse.Inventory }
+     *     
+     */
+    public GetInventoryResponse.Inventory getInventory() {
+        return inventory;
+    }
+
+    /**
+     * Sets the value of the inventory property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link GetInventoryResponse.Inventory }
+     *     
+     */
+    public void setInventory(GetInventoryResponse.Inventory value) {
+        this.inventory = value;
+    }
+
+    /**
+     * Gets the value of the state property.
+     * 
+     */
+    public int getState() {
+        return state;
+    }
+
+    /**
+     * Sets the value of the state property.
+     * 
+     */
+    public void setState(int value) {
+        this.state = value;
+    }
+
+    /**
+     * Gets the value of the timeStamp property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link String }
+     *     
+     */
+    public String getTimeStamp() {
+        return timeStamp;
+    }
+
+    /**
+     * Sets the value of the timeStamp property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *     
+     */
+    public void setTimeStamp(String value) {
+        this.timeStamp = value;
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     * 
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     * 
+     * <pre>
+     * &lt;complexType&gt;
+     *   &lt;complexContent&gt;
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+     *       &lt;sequence&gt;
+     *       &lt;/sequence&gt;
+     *     &lt;/restriction&gt;
+     *   &lt;/complexContent&gt;
+     * &lt;/complexType&gt;
+     * </pre>
+     * 
+     * 
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "")
+    public static class Inventory {
+
+
+    }
+
+}
