+++ demo/warehouse/src/main/java/com/example/warehouse/warehouse/InsertItemRequest.java
@@ -1,0 +1,84 @@
+
+package com.example.warehouse.warehouse;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Java class for anonymous complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType&gt;
+ *   &lt;complexContent&gt;
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+ *       &lt;sequence&gt;
+ *         &lt;element name="trayId" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
+ *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
+ *       &lt;/sequence&gt;
+ *     &lt;/restriction&gt;
+ *   &lt;/complexContent&gt;
+ * &lt;/complexType&gt;
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+    "trayId",
+    "name"
+})
+@XmlRootElement(name = "InsertItem", namespace = "http://example.org/warehouse")
+public class InsertItemRequest {
+
+    @XmlElement(namespace = "http://example.org/warehouse")
+    protected int trayId;
+    @XmlElement(namespace = "http://example.org/warehouse", required = true)
+    protected String name;
+
+    /**
+     * Gets the value of the trayId property.
+     * 
+     */
+    public int getTrayId() {
+        return trayId;
+    }
+
+    /**
+     * Sets the value of the trayId property.
+     * 
+     */
+    public void setTrayId(int value) {
+        this.trayId = value;
+    }
+
+    /**
+     * Gets the value of the name property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link String }
+     *     
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * Sets the value of the name property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *     
+     */
+    public void setName(String value) {
+        this.name = value;
+    }
+
+}
