+++ demo/warehouse/src/main/resources/warehouse.xsd
